feat: Add PopupManager.ts
Fixes  #3279
