+++ packages/marshaller/src/ddl2/dashboardDockPanel.ts
@@ -1,71 +1,16 @@
-import { IMonitorHandle, Widget } from "@hpcc-js/common";
+import { Widget } from "@hpcc-js/common";
 import { DockPanel, IClosable, WidgetAdapter } from "@hpcc-js/phosphor";
 import { compare } from "@hpcc-js/util";
 import { select as d3Select } from "d3-selection";
 import { Element, ElementContainer } from "./model/element";
-import { IVizPopupPanelOwner, VizChartPanel, VizPopupPanel } from "./model/vizChartPanel";
+import { IVizPopupPanelOwner } from "./model/vizChartPanel";
+import { PopupManager } from "./PopupManager";
 
 export class DashboardDockPanel extends DockPanel implements IClosable, IVizPopupPanelOwner {
 
     constructor(private _ec: ElementContainer) {
         super();
-    }
 
-    private _popups: VizPopupPanel[] = [];
-    private _popupIdx: {
-        [id: string]: {
-            panel: VizPopupPanel,
-            monitorHandle: IMonitorHandle
-        }
-    } = {};
-
-    popupPanels(): VizPopupPanel[] {
-        return this._popups;
-    }
-
-    popupChartPanels(): VizChartPanel[] {
-        return this._popups.map(p => p.widget() as VizChartPanel);
-    }
-
-    addPopup(cp: VizChartPanel) {
-        const elem = this._ec.element(cp);
-        const pp = new VizPopupPanel(this, elem)
-            .target(this.element().node())
-            .widget(cp)
-            .size({
-                width: cp.minWidth(),
-                height: cp.minHeight()
-            });
-        this._popups.push(pp);
-        this._popupIdx[cp.id()] = {
-            panel: pp,
-            monitorHandle: cp.monitor((id) => {
-                switch (id) {
-                    case "minWidth":
-                    case "minHeight":
-                        pp
-                            .resize({ width: cp.minWidth(), height: cp.minHeight() })
-                            .render()
-                            ;
-                        break;
-                }
-            })
-        };
-        this._ec.filteredBy(elem.id()).forEach(otherElem => {
-            otherElem.visualization().chartPanel().popup(pp);
-        });
-    }
-
-    removePopup(cp: VizChartPanel) {
-        const elem = this._ec.element(cp);
-        this._ec.filteredBy(elem.id()).forEach(otherElem => {
-            otherElem.visualization().chartPanel().popup(null);
-        });
-        this._popupIdx[cp.id()].panel.target(null);
-        this._popupIdx[cp.id()].monitorHandle.remove();
-        cp.target(null);
-        delete this._popupIdx[cp.id()];
-        this._popups = this._popups.filter(p => p.widget() !== cp);
     }
 
     tabTitle(element: Element): string {
@@ -140,23 +85,6 @@ export class DashboardDockPanel extends DockPanel implements IClosable, IVizPopu
         return this;
     }
 
-    syncPopups() {
-        const prevPopups = this.popupChartPanels();
-        const diffPopups = compare(prevPopups, this._ec.elements().filter(e => e.visualization().visibility() === "flyout").map((elem: Element) => elem.visualization().chartPanel()));
-
-        for (const w of diffPopups.removed) {
-            this.removePopup(w);
-        }
-
-        for (const w of diffPopups.added) {
-            this.addPopup(w);
-        }
-
-        this._popups.forEach(p => p.render());
-
-        return this;
-    }
-
     //  Events  ---
     childActivation(w: Widget, wa: WidgetAdapter) {
         super.childActivation(w, wa);
@@ -188,5 +116,8 @@ export class DashboardDockPanel extends DockPanel implements IClosable, IVizPopu
         }
         return retVal;
     }
+
+    syncPopups() {}
 }
 DashboardDockPanel.prototype._class += " marshaller_DashboardDockPanel";
+DashboardDockPanel.prototype.mixin(PopupManager);
